+++ Pulse.BL/Features/Monitors/GetMonitorsQuery.cs
@@ -1,4 +1,3 @@
 namespace Pulse.BL.Features.Monitors;
 
-public sealed record GetMonitorsQuery(MonitorStatus? Status, long? PageNumber, int? PageSize);
-
+public sealed record GetMonitorsQuery(MonitorStatus? Status, int? PageNumber, int? PageSize);
